홈,필터 xml weight 비율을 활용해서 수정함. 로그인 화면 수정함.
