+++ app/src/main/res/drawable/ic_search_20dp.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:drawable="@drawable/search"
-        android:width="24dp"
-        android:height="24dp"></item>
+        android:width="20dp"
+        android:height="20dp"></item>
 </layer-list>
