+++ app/src/main/res/layout/activity_login.xml
@@ -16,31 +16,35 @@
             android:gravity="right"
             android:orientation="horizontal">
 
-            <Button
+            <ImageButton
                 android:id="@+id/gotoMainButton"
                 style="@style/Widget.AppCompat.Button.Borderless"
-                android:layout_width="65dp"
-                android:layout_height="wrap_content"
+                android:layout_width="13dp"
+                android:layout_height="13dp"
                 android:layout_gravity="center"
                 android:layout_marginTop="21dp"
-                android:layout_marginRight="15dp"
+                android:layout_marginRight="25dp"
                 android:layout_marginBottom="15dp"
-                android:background="#00000000"
-                android:fontFamily="@font/scdream6"
-                android:minWidth="0dp"
-                android:minHeight="0dp"
-                android:text="둘러보기"
-                android:textColor="#242424" />
+                android:background="@drawable/x" />
+                //android:background="#00000000"
+
+<!--                android:minWidth="0dp"-->
+<!--                android:minHeight="0dp" -->
+
         </LinearLayout>
     </LinearLayout>
 
 
-    <ImageButton
-        android:id="@+id/userImage"
-        android:layout_width="197dp"
-        android:layout_height="181dp"
-        android:layout_gravity="center"
-        android:background="@drawable/doyouwant" />
+    <TextView
+        android:id="@+id/Login"
+        android:layout_width="match_parent"
+        android:layout_height="25dp"
+        android:layout_marginLeft="40dp"
+        android:layout_marginBottom="75dp"
+        android:fontFamily="@font/scdream6"
+        android:text="로그인"
+        android:textColor="#323232"
+        android:textSize="20sp" />
 
 
     <LinearLayout
@@ -51,7 +55,7 @@
 
         <EditText
             android:id="@+id/emailEditText"
-            android:layout_width="287dp"
+            android:layout_width="328dp"
             android:layout_height="40dp"
             android:layout_gravity="center_horizontal"
             android:layout_marginBottom="20dp"
@@ -65,7 +69,7 @@
 
         <EditText
             android:id="@+id/passwordEditText"
-            android:layout_width="287dp"
+            android:layout_width="328dp"
             android:layout_height="40dp"
             android:layout_gravity="center|center_horizontal"
             android:layout_marginLeft="40dp"
@@ -81,17 +85,17 @@
 
         <Button
             android:id="@+id/loginButton"
-            android:layout_width="280dp"
+            android:layout_width="326dp"
             android:layout_height="40dp"
             android:layout_gravity="center"
-            android:background="#86c76e"
+            android:background="#AECC6A"
             android:fontFamily="@font/scdream6"
             android:gravity="center"
             android:text="로그인"
             android:textColor="#FFFFFF" />
 
         <LinearLayout
-            android:layout_width="280dp"
+            android:layout_width="324dp"
             android:layout_height="40dp"
             android:layout_gravity="center"
             android:gravity="center"
@@ -111,7 +115,7 @@
                 android:textSize="10sp" />
 
             <View
-                android:layout_width="94dp"
+                android:layout_width="22dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="2"
                 android:background="#00000000"
